change int to long long

+++ A/fft.cpp
@@ -2,17 +2,17 @@ using cd = complex<long double>;
 const long double PI = acos(-1);
 
 void fft(vector<cd> &a, bool invert) {
-    int n = a.size(), L = __builtin_ctz(n);
-    vector<int> rev(n);
-    for (int i = 0; i < n; i++) {
+    long long n = a.size(), L = __builtin_ctz(n);
+    vector<long long> rev(n);
+    for (long long i = 0; i < n; i++) {
         rev[i] = (rev[i >> 1] | (i & 1) << L) >> 1;
         if (i < rev[i]) swap(a[i], a[rev[i]]);
     }
-    for (int len = 2; len <= n; len <<= 1) {
+    for (long long len = 2; len <= n; len <<= 1) {
         cd wlen = polar(1.0L, acos(-1.0L) / len * (invert ? -2 : 2));
-        for (int i = 0; i < n; i += len) {
+        for (long long i = 0; i < n; i += len) {
             cd w = 1;
-            for (int j = 0; j < len / 2; j++) {
+            for (long long j = 0; j < len / 2; j++) {
                 cd u = a[i + j];
                 cd v = a[i + j + len / 2] * w;
                 a[i + j] = u + v;
@@ -26,9 +26,9 @@ void fft(vector<cd> &a, bool invert) {
     }
 }
 
-vector<int> multiply(vector<int> const& a, vector<int> const& b) {
+vector<long long> multiply(vector<long long> const& a, vector<long long> const& b) {
     vector<cd> fa(a.begin(), a.end()), fb(b.begin(), b.end());
-    int n = 1;
+    long long n = 1;
     while (n < a.size() + b.size()) 
         n <<= 1;
     fa.resize(n);
@@ -36,12 +36,12 @@ vector<int> multiply(vector<int> const& a, vector<int> const& b) {
 
     fft(fa, false);
     fft(fb, false);
-    for (int i = 0; i < n; i++)
+    for (long long i = 0; i < n; i++)
         fa[i] *= fb[i];
     fft(fa, true);
 
-    vector<int> result(n);
-    for (int i = 0; i < n; i++)
+    vector<long long> result(n);
+    for (long long i = 0; i < n; i++)
         result[i] = round(fa[i].real());
     return result;
 }
